Govno

+++ TrashHackBack/src/main/java/org/example/trashhackback/repository/UserRepository.java
@@ -3,5 +3,6 @@ package org.example.trashhackback.repository;
 import org.example.trashhackback.entity.UserDao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<UserDao, Integer> {
+public interface UserRepository extends JpaRepository<UserDao, Long> {
+
 }
